Try a u32 for LineNumber

+++ src/diff/dijkstra.rs
@@ -178,7 +178,7 @@ mod tests {
     use itertools::Itertools;
     use typed_arena::Arena;
 
-    fn pos_helper(line: usize) -> Vec<SingleLineSpan> {
+    fn pos_helper(line: u32) -> Vec<SingleLineSpan> {
         vec![SingleLineSpan {
             line: line.into(),
             start_col: 0,
@@ -186,7 +186,7 @@ mod tests {
         }]
     }
 
-    fn col_helper(line: usize, col: usize) -> Vec<SingleLineSpan> {
+    fn col_helper(line: u32, col: usize) -> Vec<SingleLineSpan> {
         vec![SingleLineSpan {
             line: line.into(),
             start_col: col,
